Messages exchanged between main window and an iframe must be tagged with window_id - in order to identify a proper iframe.
Due to Opera bug, it's impossible to check where did the event (onmessage) came from - "event.source === iframe...window" fails.
As a result, if there are more iframes open the communication gets completely messed up. To avoid that, we need to be able to identify a particular iframe from main window. The problem is how to share a handle with an iframe - it's on a foreign domain. For that we use an old hack - we push the handle as uri fragment (#hash). That way, the iframe.html document may still be cached and the iframe can be properly identified on both sides.