Enable chunking-detection for streaming protocols (with the exception of websockets).
It's possible that some users will be behind a broken proxy that doesn't allow chunking protocols. Instead of handling this problem in every streaming protocol manually (like previously we did for xhr-polling/streaming), we will run a single test that ought to detect if chunking works at all. The underlying assumption is that if chunking works for XHR protocol, it is likely that it will work for htmlfile and eventsource.