Chunking test shall work on opera.
Although opera chunking handling in xhr is, quite bad - we can't do xhr-streaming, we can indeed use it for chunking detection. Opera limitation is that it runs onreadystate=3 event exactly once. This is enough for chunking detecion.

If we get readystate == 3, and the contents suggest that we haven't received full response - clearly chunking works.