Remove max parameter from util.random_string
The only place that uses it, trans-receiver-htmlfile.js, doesn't need it; it
prepends 'a' to the front of the identifier, so numbers aren't a problem. This
is a nuisance for making util.random_string use a real random number generator;
values of max that aren't powers of two would bias towards some values.

That caller doesn't really care anyway, but we may as well remove the
unnecessary parameter.