Allow custom session ids
Every time a new connection is made to a sockjs server, a session id is
generated in the client and sent to the server. This session id helps
the server distinguish connections. By default, the current code
generates a random 8-character-long string as session id. There is
currently no way to increase the entropy or even generate your own
session ids, so this commit is an attempt to tackle that limitation by
specifying the option called sessionId in the options passed to SockJS.